Downgraded Petsc version!!
This is an important commit as it may (will) break your installation. Therefore we will update the version number of CDMATH to 0.2.
This version requires the use of Petsc 3.4, released in May 2013, and not Petsc 3.5 anymore. The reason is mainly that Petsc 3.4 is available in the repositories of major distributions such as Ubuntu 14.04 or Ubuntu 14.10.
We hope, as a consequence, that CDMATH will be much easier to install than before. This is consistent with our aim of making CDMATH an accessible toolbox, even for numericists and thermal-hydraulicists who do not wish to have extented knowledge about playing with dependencies. We know from feedback that it will help at least a few of our users.
Do not hesitate to get in touch with us, the development team, if you have feedback.